Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/malformed-HTML-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/malformed-HTML.html&q=%3Ca%3Cimg/src/onerror=alert(1)//%3C' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.


--------
Frame: '<!--frame1-->'
--------
